requirements.txt
feedparser
beautifulsoup4
python-dotenv